+++ src/apps/Query/index.ts
@@ -1,7 +1,42 @@
-import { ClientApi } from '@equinor/lighthouse-portal-client';
+import { ClientApi, httpClient } from '@equinor/lighthouse-portal-client';
+import { Query } from './types';
+import {
+    analyticsConfig,
+    tableConfig,
+    filterConfig,
+    statusBarConfig,
+    gardenConfig,
+    sidesheetConfig,
+} from './utility/config';
+import { sortQueryByStatus } from './utility/helpers';
+
+async function responseAsync(signal?: AbortSignal | undefined): Promise<Response> {
+    const { FAM } = httpClient();
+    return await FAM.post('v0.1/dynamic/completion/custom_query/JCA', {
+        body: JSON.stringify({}),
+        signal,
+    });
+}
+
+async function responseParser(response: Response) {
+    const parsedResponse = JSON.parse(await response.text()) as Query[];
+    return parsedResponse.sort(sortQueryByStatus);
+}
 
 export function setup(appApi: ClientApi): void {
-    appApi.createPowerBiViewer().registerFusionPowerBi({
-        reportURI: 'pp-query-analytics',
-    });
+    appApi
+        .createWorkSpace<Query>({
+            objectIdentifier: 'queryId',
+            customSidesheetOptions: sidesheetConfig('WorkspaceSideSheet'),
+        })
+        .registerDataSource({ responseAsync: responseAsync, responseParser: responseParser })
+        .registerTableOptions(tableConfig)
+        .registerFilterOptions(filterConfig)
+        .registerStatusItems(statusBarConfig)
+        .registerGardenOptions(gardenConfig)
+        .registerPowerBIOptions(analyticsConfig)
+        .registerSearchOptions([
+            { name: 'Query No', valueFormatter: (query) => query.queryNo },
+            { name: 'Title', valueFormatter: (query) => query.title },
+        ]);
 }