+++ src/apps/ScopeChangeRequest/Components/Workflow/Criteria/Components/CriteriaActionOverlay.tsx
@@ -1,10 +1,9 @@
 import { useAtom } from '@dbeining/react-atom';
 import { Button } from '@equinor/eds-core-react';
-import { SignWithComment } from './SignWithComment/SignWithComment';
 import { useScopeChangeContext } from '../../../../hooks/context/useScopeChangeContext';
-import { actionWithCommentAtom, resetSigningAtom } from '../../Atoms/signingAtom';
 import { ReassignBar } from '../../ReassignBar/ReassignBar';
 import styled from 'styled-components';
+import { actionWithCommentAtom, resetSigningAtom } from '@equinor/Workflow';
 
 export const CriteriaActionOverlay = (): JSX.Element | null => {
     const state = useAtom(actionWithCommentAtom);
@@ -14,27 +13,16 @@ export const CriteriaActionOverlay = (): JSX.Element | null => {
         return null;
     }
     return (
-        <>
-            {state.action === 'Reassign' ? (
-                <ButtonContainer>
-                    <ReassignBar
-                        criteriaId={state.criteriaId}
-                        requestId={requestId}
-                        stepId={state.stepId}
-                    />
-                    <Button variant="outlined" onClick={resetSigningAtom}>
-                        Cancel
-                    </Button>
-                </ButtonContainer>
-            ) : (
-                <SignWithComment
-                    action={state.action}
-                    buttonText={state.buttonText}
-                    criteriaId={state.criteriaId}
-                    stepId={state.stepId}
-                />
-            )}
-        </>
+        <ButtonContainer>
+            <ReassignBar
+                criteriaId={state.criteriaId}
+                requestId={requestId}
+                stepId={state.stepId}
+            />
+            <Button variant="outlined" onClick={resetSigningAtom}>
+                Cancel
+            </Button>
+        </ButtonContainer>
     );
 };