+++ src/apps/ScopeChangeRequest/Components/Workflow/Criteria/Components/SignWithComment/signWithComment.styles.ts
@@ -1,8 +1,0 @@
-import styled from 'styled-components';
-
-export const ButtonsContainer = styled.div`
-    gap: 5px;
-    display: flex;
-    width: 100%;
-    justify-content: flex-end;
-`;