+++ src/apps/apps.ts
@@ -643,30 +643,17 @@ export const apps: AppManifest[] = [
         appEnv: 'prod',
     },
     {
-        title: 'Query analytics',
-        shortName: 'query-analytics',
+        title: 'Query',
+        shortName: 'query-workspace',
         color: '#0364B8',
         groupe: Apps.Collaboration,
         icon: '',
+        tags: [],
         app: {
-            appType: 'PowerBIViewer',
+            appType: 'Workspace',
             setup: querySetup,
         },
-        tags: ['PowerBI'],
         appEnv: 'prod',
-    },
-    {
-        title: 'Queries',
-        shortName: 'queries',
-        color: '#0364B8',
-        groupe: Apps.Collaboration,
-        icon: '',
-        uri: (isProduction: boolean) =>
-            `https://${
-                isProduction ? 'procosys' : 'procosystest'
-            }.equinor.com/JOHAN_CASTBERG/Search?searchType=Query`,
-        tags: ['link', 'procosys'],
-        appEnv: 'test',
     },
     {
         title: 'AKSO procedures',